+++ test/Test/HspecSpec.hs
@@ -23,10 +23,11 @@ spec = do
         H.it "foo" H.pending
       r `shouldSatisfy` any (== "     # PENDING: No reason given")
 
-    it "accepts an optional message, which is included in the report" $ do
+  describe "pending" $ do
+    it "specifies a pending example with a reason for why it's pending" $ do
       r <- runSpec $ do
         H.it "foo" $ do
-          H.pending "for some reason"
+          H.pendingWith "for some reason"
       r `shouldSatisfy` any (== "     # PENDING: for some reason")
 
   describe "describe" $ do